fastq-registration/requirements.txt
boto3==1.34
requests==2.32.3
python-dotenv==1.0
biopython==1.83
pydantic==2.7